Homework 3 fix
